# Apply micro-optimisation for Table & Trip APIs