<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1096.35<br/>plain u32: 1101.89<br/>aliased double: 965.261<br/>plain double: 963.328 | aliased u32: 1099.62<br/>plain u32: 1104.85<br/>aliased double: 963.952<br/>plain double: 957.935 |
| e2e_match_ch | Total: 2903.632402420044ms<br/>Min time: 2.317190170288086ms<br/>Mean time: 22.165132842901098ms<br/>Median time: 15.163898468017578ms<br/>95th percentile: 69.90289688110352ms<br/>99th percentile: 84.40463542938225ms<br/>Max time: 97.20349311828613ms | Total: 2903.6059379577637ms<br/>Min time: 2.390623092651367ms<br/>Mean time: 22.164930824105067ms<br/>Median time: 14.85753059387207ms<br/>95th percentile: 72.19231128692627ms<br/>99th percentile: 86.08164787292475ms<br/>Max time: 99.8220443725586ms |
| e2e_match_mld | Total: 2058.0296516418457ms<br/>Min time: 2.1736621856689453ms<br/>Mean time: 15.710150012533173ms<br/>Median time: 8.475303649902344ms<br/>95th percentile: 49.976468086242676ms<br/>99th percentile: 60.28988361358636ms<br/>Max time: 69.13518905639648ms | Total: 2021.5742588043213ms<br/>Min time: 2.0301342010498047ms<br/>Mean time: 15.431864571025354ms<br/>Median time: 8.957386016845703ms<br/>95th percentile: 49.45647716522217ms<br/>99th percentile: 58.389115333557065ms<br/>Max time: 66.33543968200684ms |
| e2e_nearest_ch | Total: 1338.7506008148193ms<br/>Min time: 1.1363029479980469ms<br/>Mean time: 1.3387506008148193ms<br/>Median time: 1.2537240982055664ms<br/>95th percentile: 1.750481128692627ms<br/>99th percentile: 1.8105530738830566ms<br/>Max time: 2.0513534545898438ms | Total: 1366.973876953125ms<br/>Min time: 1.1582374572753906ms<br/>Mean time: 1.366973876953125ms<br/>Median time: 1.2753009796142578ms<br/>95th percentile: 1.774132251739502ms<br/>99th percentile: 1.8685674667358398ms<br/>Max time: 2.1600723266601562ms |
| e2e_nearest_mld | Total: 1355.5173873901367ms<br/>Min time: 1.1615753173828125ms<br/>Mean time: 1.3555173873901367ms<br/>Median time: 1.267075538635254ms<br/>95th percentile: 1.7766952514648438ms<br/>99th percentile: 1.8379759788513184ms<br/>Max time: 1.9915103912353516ms | Total: 1340.7692909240723ms<br/>Min time: 1.1463165283203125ms<br/>Mean time: 1.3407692909240723ms<br/>Median time: 1.2546777725219727ms<br/>95th percentile: 1.7500042915344238ms<br/>99th percentile: 1.8103218078613281ms<br/>Max time: 1.8572807312011719ms |
| e2e_route_ch | Total: 3120.9702491760254ms<br/>Min time: 1.401662826538086ms<br/>Mean time: 3.1209702491760254ms<br/>Median time: 3.1474828720092773ms<br/>95th percentile: 4.100096225738525ms<br/>99th percentile: 4.490950107574462ms<br/>Max time: 5.212306976318359ms | Total: 3061.6183280944824ms<br/>Min time: 1.3391971588134766ms<br/>Mean time: 3.0616183280944824ms<br/>Median time: 3.082156181335449ms<br/>95th percentile: 4.024291038513184ms<br/>99th percentile: 4.383678436279297ms<br/>Max time: 5.019903182983398ms |
| e2e_route_mld | Total: 3712.446928024292ms<br/>Min time: 1.3332366943359375ms<br/>Mean time: 3.712446928024292ms<br/>Median time: 3.743290901184082ms<br/>95th percentile: 5.0673484802246085ms<br/>99th percentile: 5.528898239135741ms<br/>Max time: 6.256580352783203ms | Total: 3606.574058532715ms<br/>Min time: 1.318216323852539ms<br/>Mean time: 3.606574058532715ms<br/>Median time: 3.641486167907715ms<br/>95th percentile: 4.9071311950683585ms<br/>99th percentile: 5.389270782470703ms<br/>Max time: 6.054878234863281ms |
| e2e_table_ch | Total: 15904.053449630737ms<br/>Min time: 2.0303726196289062ms<br/>Mean time: 15.904053449630737ms<br/>Median time: 15.166044235229492ms<br/>95th percentile: 29.210782051086422ms<br/>99th percentile: 30.741758346557617ms<br/>Max time: 34.508466720581055ms | Total: 16110.6436252594ms<br/>Min time: 1.9288063049316406ms<br/>Mean time: 16.1106436252594ms<br/>Median time: 15.396952629089355ms<br/>95th percentile: 29.858839511871338ms<br/>99th percentile: 31.190974712371826ms<br/>Max time: 33.14042091369629ms |
| e2e_table_mld | Total: 63630.54847717285ms<br/>Min time: 4.227876663208008ms<br/>Mean time: 63.63054847717285ms<br/>Median time: 60.47976016998291ms<br/>95th percentile: 122.23318815231323ms<br/>99th percentile: 130.2273178100586ms<br/>Max time: 133.40020179748535ms | Total: 63228.24573516846ms<br/>Min time: 4.151105880737305ms<br/>Mean time: 63.22824573516846ms<br/>Median time: 60.08493900299072ms<br/>95th percentile: 121.36526107788086ms<br/>99th percentile: 128.21901082992554ms<br/>Max time: 131.4857006072998ms |
| e2e_trip_ch | Total: 10733.873844146729ms<br/>Min time: 1.5108585357666016ms<br/>Mean time: 10.733873844146729ms<br/>Median time: 10.247230529785156ms<br/>95th percentile: 18.78179311752319ms<br/>99th percentile: 20.58562994003296ms<br/>Max time: 21.63553237915039ms | Total: 10768.372058868408ms<br/>Min time: 1.5866756439208984ms<br/>Mean time: 10.768372058868408ms<br/>Median time: 10.27679443359375ms<br/>95th percentile: 18.883514404296875ms<br/>99th percentile: 20.550918579101562ms<br/>Max time: 21.411418914794922ms |
| e2e_trip_mld | Total: 18036.177396774292ms<br/>Min time: 1.978158950805664ms<br/>Mean time: 18.036177396774292ms<br/>Median time: 17.68338680267334ms<br/>95th percentile: 29.32676076889038ms<br/>99th percentile: 30.93783140182495ms<br/>Max time: 33.60414505004883ms | Total: 18094.83575820923ms<br/>Min time: 1.6837120056152344ms<br/>Mean time: 18.09483575820923ms<br/>Median time: 17.726421356201172ms<br/>95th percentile: 29.52185869216919ms<br/>99th percentile: 31.005425453186035ms<br/>Max time: 32.550811767578125ms |
| json-render | String: 6.78858ms<br/>Stringstream: 8.80808ms<br/>Vector: 6.90945ms | String: 6.65951ms<br/>Stringstream: 9.30541ms<br/>Vector: 6.92236ms |
| match_ch | Default radius: <br/>4.41918ms/req at 82 coordinate<br/>0.0538924ms/coordinate<br/>Radius 5m: <br/>4.39787ms/req at 82 coordinate<br/>0.0536326ms/coordinate<br/>Radius 10m: <br/>15.0352ms/req at 82 coordinate<br/>0.183356ms/coordinate<br/>Radius 15m: <br/>36.6981ms/req at 82 coordinate<br/>0.447538ms/coordinate<br/>Radius 30m: <br/>312.319ms/req at 82 coordinate<br/>3.80877ms/coordinate | Default radius: <br/>4.4231ms/req at 82 coordinate<br/>0.0539403ms/coordinate<br/>Radius 5m: <br/>4.41239ms/req at 82 coordinate<br/>0.0538096ms/coordinate<br/>Radius 10m: <br/>15.0203ms/req at 82 coordinate<br/>0.183175ms/coordinate<br/>Radius 15m: <br/>36.7226ms/req at 82 coordinate<br/>0.447837ms/coordinate<br/>Radius 30m: <br/>313.396ms/req at 82 coordinate<br/>3.82191ms/coordinate |
| match_mld | Default radius: <br/>2.75999ms/req at 82 coordinate<br/>0.0336584ms/coordinate<br/>Radius 5m: <br/>2.83163ms/req at 82 coordinate<br/>0.0345321ms/coordinate<br/>Radius 10m: <br/>10.2675ms/req at 82 coordinate<br/>0.125213ms/coordinate<br/>Radius 15m: <br/>25.8417ms/req at 82 coordinate<br/>0.315142ms/coordinate<br/>Radius 30m: <br/>303.066ms/req at 82 coordinate<br/>3.69593ms/coordinate | Default radius: <br/>2.76324ms/req at 82 coordinate<br/>0.033698ms/coordinate<br/>Radius 5m: <br/>2.76415ms/req at 82 coordinate<br/>0.0337092ms/coordinate<br/>Radius 10m: <br/>10.1465ms/req at 82 coordinate<br/>0.123738ms/coordinate<br/>Radius 15m: <br/>26.0569ms/req at 82 coordinate<br/>0.317767ms/coordinate<br/>Radius 30m: <br/>305.527ms/req at 82 coordinate<br/>3.72594ms/coordinate |
| osrm_contract | Time: 93.59s Peak RAM: 185.73MB | Time: 92.57s Peak RAM: 185.73MB |
| osrm_customize | Time: 1.31s Peak RAM: 115.01MB | Time: 1.31s Peak RAM: 115.00MB |
| osrm_extract | Time: 12.28s Peak RAM: 436.53MB | Time: 12.26s Peak RAM: 423.04MB |
| osrm_partition | Time: 2.22s Peak RAM: 148.80MB | Time: 2.20s Peak RAM: 148.91MB |
| packedvector | random write:<br/>std::vector 9854.45 ms<br/>util::packed_vector 82087.1 ms<br/>slowdown: 8.32996<br/>random read:<br/>std::vector 8557.81 ms<br/>util::packed_vector 33708 ms<br/>slowdown: 3.93886 | random write:<br/>std::vector 9922.7 ms<br/>util::packed_vector 78600.8 ms<br/>slowdown: 7.92131<br/>random read:<br/>std::vector 8549.4 ms<br/>util::packed_vector 31463.8 ms<br/>slowdown: 3.68024 |
| random_match_ch | 1000 matches, default radius<br/>total: 6967.45ms<br/>avg: 6.97ms<br/>min: 0.00ms<br/>max: 484.31ms<br/>p99: 109.95ms<br/><br/>1000 matches, radius=10<br/>total: 34795.65ms<br/>avg: 34.80ms<br/>min: 0.00ms<br/>max: 1878.00ms<br/>p99: 1845.85ms<br/><br/>1000 matches, radius=20<br/>total: 68408.35ms<br/>avg: 68.41ms<br/>min: 0.00ms<br/>max: 9270.68ms<br/>p99: 1237.93ms | 1000 matches, default radius<br/>total: 6905.24ms<br/>avg: 6.91ms<br/>min: 0.00ms<br/>max: 462.19ms<br/>p99: 109.61ms<br/><br/>1000 matches, radius=10<br/>total: 34304.16ms<br/>avg: 34.30ms<br/>min: 0.00ms<br/>max: 1819.53ms<br/>p99: 1799.55ms<br/><br/>1000 matches, radius=20<br/>total: 67118.23ms<br/>avg: 67.12ms<br/>min: 0.00ms<br/>max: 8864.83ms<br/>p99: 1235.73ms |
| random_match_mld | 1000 matches, default radius<br/>total: 5133.96ms<br/>avg: 5.13ms<br/>min: 0.00ms<br/>max: 382.72ms<br/>p99: 69.61ms<br/><br/>1000 matches, radius=10<br/>total: 26687.29ms<br/>avg: 26.69ms<br/>min: 0.00ms<br/>max: 1554.12ms<br/>p99: 1531.76ms<br/><br/>1000 matches, radius=20<br/>total: 52386.10ms<br/>avg: 52.39ms<br/>min: 0.00ms<br/>max: 6960.01ms<br/>p99: 789.28ms | 1000 matches, default radius<br/>total: 5217.87ms<br/>avg: 5.22ms<br/>min: 0.00ms<br/>max: 393.21ms<br/>p99: 70.08ms<br/><br/>1000 matches, radius=10<br/>total: 26903.96ms<br/>avg: 26.90ms<br/>min: 0.00ms<br/>max: 1574.04ms<br/>p99: 1548.24ms<br/><br/>1000 matches, radius=20<br/>total: 53220.44ms<br/>avg: 53.22ms<br/>min: 0.00ms<br/>max: 7074.59ms<br/>p99: 795.70ms |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>total: 415.67ms<br/>avg: 0.04ms<br/>min: 0.01ms<br/>max: 0.22ms<br/>p99: 0.10ms<br/><br/>10000 nearest, number_of_results=5<br/>total: 572.54ms<br/>avg: 0.06ms<br/>min: 0.02ms<br/>max: 0.15ms<br/>p99: 0.12ms<br/><br/>10000 nearest, number_of_results=10<br/>total: 743.11ms<br/>avg: 0.07ms<br/>min: 0.03ms<br/>max: 0.22ms<br/>p99: 0.14ms | 10000 nearest, number_of_results=1<br/>total: 411.08ms<br/>avg: 0.04ms<br/>min: 0.01ms<br/>max: 0.19ms<br/>p99: 0.10ms<br/><br/>10000 nearest, number_of_results=5<br/>total: 567.10ms<br/>avg: 0.06ms<br/>min: 0.02ms<br/>max: 0.15ms<br/>p99: 0.12ms<br/><br/>10000 nearest, number_of_results=10<br/>total: 732.07ms<br/>avg: 0.07ms<br/>min: 0.03ms<br/>max: 0.18ms<br/>p99: 0.13ms |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>total: 423.13ms<br/>avg: 0.04ms<br/>min: 0.01ms<br/>max: 0.23ms<br/>p99: 0.11ms<br/><br/>10000 nearest, number_of_results=5<br/>total: 581.51ms<br/>avg: 0.06ms<br/>min: 0.02ms<br/>max: 0.19ms<br/>p99: 0.13ms<br/><br/>10000 nearest, number_of_results=10<br/>total: 733.39ms<br/>avg: 0.07ms<br/>min: 0.03ms<br/>max: 0.17ms<br/>p99: 0.14ms | 10000 nearest, number_of_results=1<br/>total: 413.25ms<br/>avg: 0.04ms<br/>min: 0.01ms<br/>max: 0.22ms<br/>p99: 0.10ms<br/><br/>10000 nearest, number_of_results=5<br/>total: 565.64ms<br/>avg: 0.06ms<br/>min: 0.02ms<br/>max: 0.15ms<br/>p99: 0.12ms<br/><br/>10000 nearest, number_of_results=10<br/>total: 729.46ms<br/>avg: 0.07ms<br/>min: 0.03ms<br/>max: 0.17ms<br/>p99: 0.13ms |
| random_route_ch | 10000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>total: 21023.48ms<br/>avg: 2.10ms<br/>min: 0.14ms<br/>max: 3.87ms<br/>p99: 3.19ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>total: 9588.59ms<br/>avg: 0.96ms<br/>min: 0.07ms<br/>max: 2.15ms<br/>p99: 1.56ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>total: 18418.63ms<br/>avg: 1.84ms<br/>min: 0.06ms<br/>max: 4.78ms<br/>p99: 4.10ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>total: 10024.37ms<br/>avg: 1.00ms<br/>min: 0.08ms<br/>max: 2.03ms<br/>p99: 1.46ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>total: 4021.14ms<br/>avg: 0.40ms<br/>min: 0.05ms<br/>max: 0.76ms<br/>p99: 0.60ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>total: 9648.26ms<br/>avg: 0.96ms<br/>min: 0.06ms<br/>max: 3.20ms<br/>p99: 2.38ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 599.35ms<br/>avg: 0.06ms<br/>min: 0.01ms<br/>max: 1.69ms<br/>p99: 0.68ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 625.03ms<br/>avg: 0.06ms<br/>min: 0.01ms<br/>max: 0.57ms<br/>p99: 0.42ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>total: 853.75ms<br/>avg: 0.09ms<br/>min: 0.01ms<br/>max: 1.83ms<br/>p99: 1.17ms | 10000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>total: 20827.68ms<br/>avg: 2.08ms<br/>min: 0.11ms<br/>max: 3.94ms<br/>p99: 3.07ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>total: 9654.16ms<br/>avg: 0.97ms<br/>min: 0.08ms<br/>max: 2.21ms<br/>p99: 1.57ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>total: 18397.43ms<br/>avg: 1.84ms<br/>min: 0.06ms<br/>max: 5.00ms<br/>p99: 3.85ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>total: 10115.71ms<br/>avg: 1.01ms<br/>min: 0.08ms<br/>max: 1.97ms<br/>p99: 1.43ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>total: 4074.67ms<br/>avg: 0.41ms<br/>min: 0.05ms<br/>max: 0.72ms<br/>p99: 0.60ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>total: 9260.27ms<br/>avg: 0.93ms<br/>min: 0.06ms<br/>max: 4.28ms<br/>p99: 2.03ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 605.01ms<br/>avg: 0.06ms<br/>min: 0.01ms<br/>max: 1.66ms<br/>p99: 0.73ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 633.12ms<br/>avg: 0.06ms<br/>min: 0.01ms<br/>max: 0.63ms<br/>p99: 0.42ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>total: 855.91ms<br/>avg: 0.09ms<br/>min: 0.01ms<br/>max: 1.98ms<br/>p99: 1.18ms |
| random_route_mld | 10000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>total: 40783.23ms<br/>avg: 4.08ms<br/>min: 0.14ms<br/>max: 9.78ms<br/>p99: 7.04ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>total: 14852.09ms<br/>avg: 1.49ms<br/>min: 0.07ms<br/>max: 3.06ms<br/>p99: 2.63ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>total: 41668.86ms<br/>avg: 4.17ms<br/>min: 0.06ms<br/>max: 10.26ms<br/>p99: 8.43ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>total: 29863.21ms<br/>avg: 2.99ms<br/>min: 0.08ms<br/>max: 9.25ms<br/>p99: 5.27ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>total: 8999.41ms<br/>avg: 0.90ms<br/>min: 0.04ms<br/>max: 3.10ms<br/>p99: 1.61ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>total: 32845.94ms<br/>avg: 3.28ms<br/>min: 0.05ms<br/>max: 8.17ms<br/>p99: 6.74ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 782.94ms<br/>avg: 0.08ms<br/>min: 0.01ms<br/>max: 4.40ms<br/>p99: 1.55ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 870.73ms<br/>avg: 0.09ms<br/>min: 0.01ms<br/>max: 1.40ms<br/>p99: 1.02ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>total: 1658.06ms<br/>avg: 0.17ms<br/>min: 0.01ms<br/>max: 5.08ms<br/>p99: 3.55ms | 10000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>total: 41020.83ms<br/>avg: 4.10ms<br/>min: 0.13ms<br/>max: 10.07ms<br/>p99: 7.07ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>total: 14638.66ms<br/>avg: 1.46ms<br/>min: 0.07ms<br/>max: 2.89ms<br/>p99: 2.58ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>total: 42755.51ms<br/>avg: 4.28ms<br/>min: 0.06ms<br/>max: 10.66ms<br/>p99: 8.82ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>total: 29621.45ms<br/>avg: 2.96ms<br/>min: 0.08ms<br/>max: 9.25ms<br/>p99: 5.25ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>total: 9454.40ms<br/>avg: 0.95ms<br/>min: 0.04ms<br/>max: 2.17ms<br/>p99: 1.73ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>total: 32372.99ms<br/>avg: 3.24ms<br/>min: 0.05ms<br/>max: 9.59ms<br/>p99: 6.54ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 782.63ms<br/>avg: 0.08ms<br/>min: 0.01ms<br/>max: 4.41ms<br/>p99: 1.60ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 861.62ms<br/>avg: 0.09ms<br/>min: 0.01ms<br/>max: 1.28ms<br/>p99: 1.00ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>total: 1646.23ms<br/>avg: 0.16ms<br/>min: 0.01ms<br/>max: 4.99ms<br/>p99: 3.50ms |
| random_table_ch | 250 tables, 3 coordinates<br/>total: 176.75ms<br/>avg: 0.71ms<br/>min: 0.50ms<br/>max: 1.85ms<br/>p99: 1.05ms<br/><br/>250 tables, 25 coordinates<br/>total: 1451.02ms<br/>avg: 5.80ms<br/>min: 5.03ms<br/>max: 6.87ms<br/>p99: 6.48ms<br/><br/>250 tables, 50 coordinates<br/>total: 2902.47ms<br/>avg: 11.61ms<br/>min: 10.68ms<br/>max: 13.03ms<br/>p99: 12.41ms<br/><br/>250 tables, 100 coordinates<br/>total: 6206.47ms<br/>avg: 24.83ms<br/>min: 23.29ms<br/>max: 26.46ms<br/>p99: 26.07ms | 250 tables, 3 coordinates<br/>total: 179.65ms<br/>avg: 0.72ms<br/>min: 0.52ms<br/>max: 1.66ms<br/>p99: 1.01ms<br/><br/>250 tables, 25 coordinates<br/>total: 1475.57ms<br/>avg: 5.90ms<br/>min: 5.21ms<br/>max: 6.39ms<br/>p99: 6.36ms<br/><br/>250 tables, 50 coordinates<br/>total: 2991.92ms<br/>avg: 11.97ms<br/>min: 10.91ms<br/>max: 12.94ms<br/>p99: 12.74ms<br/><br/>250 tables, 100 coordinates<br/>total: 6411.47ms<br/>avg: 25.65ms<br/>min: 24.33ms<br/>max: 27.04ms<br/>p99: 26.84ms |
| random_table_mld | 250 tables, 3 coordinates<br/>total: 733.85ms<br/>avg: 2.94ms<br/>min: 2.34ms<br/>max: 4.01ms<br/>p99: 3.90ms<br/><br/>250 tables, 25 coordinates<br/>total: 6837.54ms<br/>avg: 27.35ms<br/>min: 24.68ms<br/>max: 31.78ms<br/>p99: 30.28ms<br/><br/>250 tables, 50 coordinates<br/>total: 14364.12ms<br/>avg: 57.46ms<br/>min: 53.29ms<br/>max: 65.44ms<br/>p99: 61.66ms<br/><br/>250 tables, 100 coordinates<br/>total: 31498.12ms<br/>avg: 125.99ms<br/>min: 120.81ms<br/>max: 138.52ms<br/>p99: 132.78ms | 250 tables, 3 coordinates<br/>total: 733.94ms<br/>avg: 2.94ms<br/>min: 2.32ms<br/>max: 3.99ms<br/>p99: 3.90ms<br/><br/>250 tables, 25 coordinates<br/>total: 6941.04ms<br/>avg: 27.76ms<br/>min: 25.31ms<br/>max: 32.02ms<br/>p99: 30.93ms<br/><br/>250 tables, 50 coordinates<br/>total: 14866.65ms<br/>avg: 59.47ms<br/>min: 54.83ms<br/>max: 63.56ms<br/>p99: 63.43ms<br/><br/>250 tables, 100 coordinates<br/>total: 31379.02ms<br/>avg: 125.52ms<br/>min: 119.84ms<br/>max: 132.03ms<br/>p99: 131.54ms |
…